// source/frontend/src/models/api/generated/ApiGen_Requests_FileDownloadResponse.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */

// LINK: @backend/apimodels/Models/Requests/Http/FileDownloadResponse.cs
export interface ApiGen_Requests_FileDownloadResponse {
  filePayload: string | null;
  size: number;
  fileName: string | null;
  fileNameExtension: string | null;
  fileNameWithoutExtension: string | null;
  mimetype: string | null;
  encodingType: string | null;
}